+++ LocalPackages/net.peeweek.gameplay-ingredients/Runtime/LevelScripting/Actions/AudioPlayClipAction.cs
@@ -1,8 +1,5 @@
-using System.Collections;
-using System;
 using UnityEngine;
-using UnityEngine.Audio;
-using UnityEngine.Events;
+using NaughtyAttributes;
 
 namespace GameplayIngredients.Actions
 {
@@ -10,6 +7,12 @@ namespace GameplayIngredients.Actions
     {
         public AudioClip Clip;
         public AudioSource Source;
+        public bool RandomizePitch = false;
+        [ShowIf("RandomizePitch")]
+        public Vector2 PitchRange = new Vector2(0,3); 
+        public bool RandomizeVolume = false;
+        [ShowIf("RandomizeVolume")]
+        public Vector2 VolumeRange = new Vector2(0, 1);
 
         public override void Execute(GameObject instigator = null)
         {
@@ -17,12 +20,15 @@ namespace GameplayIngredients.Actions
             {
                 Source.Stop();
 
+                if (RandomizePitch)
+                    Source.pitch = Random.Range(PitchRange.x, PitchRange.y);
+                if (RandomizeVolume)
+                    Source.volume = Random.Range(VolumeRange.x, VolumeRange.y);
                 if (Clip != null)
                     Source.clip = Clip;
 
                 Source.Play();
             }
         }
-
     }
 }