+++ LocalPackages/net.peeweek.gameplay-ingredients/CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## 2019.3.3
+
+#### Added
+
+- Added Counters : Integer counters stored in the scene
+- Added CounterAction and CounterLogic
+- Added Create State button to StateMachines
+- Added State Machine templates in the GameObject Creation Menu
+
+#### Changed
+
+- Messager now send messages through a copy of its source listeners array instead of the actual enumerator. This prevents from modifying the call array at the same frame if a OnMessageEvent becomes disabled during execution.
+- Changed SceneViewToolbar Accessibility (now public to give access to OnSceneViewToolbarGUI)
+
+#### Fixed
+
+- Fixes in changelog (bad changes in 2019.3.2)
+
+## 2019.3.2
+
+#### Changed
+
+- **Delayed Logic** now has a mode that allows you to have a random delay within a range.
+- **Audio Play Clip Action** now allows you to randomize volume and pitch within a range of values every time you play the clip.
+
+#### Fixed
+
+- Fixed selection of Call tree window that became broken somehow in 2019.3.0f1
+
 ## 2019.3.1
 
 #### Changed
@@ -13,6 +42,8 @@
 * **Call Tree Explorer**: Added Category for Erroneous Calls
 * Added **ToggleBehaviourAction** working the same as ToggleGameObjectAction, but for behaviour components instead.
 * **SendMessageBehaviour** (Timeline Send Message Tracks/Clips) now displays an instigator game object field in the inspector to be attached to the sent message.
+* Added **VFXSetPropertyAction**
+* Added **VFXSendEventAction**
 
 #### Fixed